refactor(nocode-starter-backend): refactor project file names
